Add flex style to exercise 04
implement all requirement for exercise 04 flex
but its a little bit different. In my case,
the flex item is wrapped again with flex
and then change the flex direction to column.
in the requirement, the flex direction still
horizontal, then use wrap on it.
then i just realize that the solution is way more
simple